Save a new musicAlbum and genre in json files

+++ file_manager.rb
@@ -3,8 +3,10 @@ require 'json'
 
 def createFiles(file_names)
     file_names.each do |file_name| 
+       if File.exist?("./json/#{file_name}.json") == false
         FileUtils.touch("./json/#{file_name}.json")
         File.write("./json/#{file_name}.json", JSON.dump([]))
+       end
     end
 end
 
@@ -14,6 +16,6 @@ def save_file(file_name, content)
 end
 
 def read_file(file_name)
-    JSON.parse(File.read("./json/#{file_name}.json"))
+   return JSON.parse(File.read("./json/#{file_name}.json"))
 end
